+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [1.3.0] - 2025-02-03
 
 ### Added
 - auth:
@@ -122,7 +122,8 @@ and this project adheres to
 
 ## [1.0.0] - 2024-04-08
 
-[unreleased]: https://github.com/rackslab/RFL/compare/v1.2.0...HEAD
+[unreleased]: https://github.com/rackslab/RFL/compare/v1.3.0...HEAD
+[1.3.0]: https://github.com/rackslab/RFL/releases/tag/v1.3.0
 [1.2.0]: https://github.com/rackslab/RFL/releases/tag/v1.2.0
 [1.1.1]: https://github.com/rackslab/RFL/releases/tag/v1.1.1
 [1.1.0]: https://github.com/rackslab/RFL/releases/tag/v1.1.0